fix: 逐文件审查修复 - dead code清理、硬编码i18n、防抖逻辑、i18n补齐
代码审查逐文件检查并修复8个UI文件：

1. image_viewer.dart - 删除dead code（viewer_page已有自己的_ImageViewer实现）
2. audio_player_widget.dart:
   - 修复硬编码中文 '音频播放器初始化失败' (改用英文debugPrint)
   - 修复硬编码中文 '未知音频' (改用文件名fallback)
   - 修复重复赋值 _duration = _player.duration (删除第71行)
3. video_player_widget.dart - 修复clamp(1,int)返回num导致double赋值报错
4. search_bar.dart:
   - 删除多余空行
   - 修复防抖逻辑Future.delayed未取消旧timer（改用Timer+cancel+dispose）
   - 添加dart:async导入
5. file_browser_dialog.dart:
   - _showHidden 从 final 改为 mutable
   - 添加隐藏文件切换按钮(visibility/visibility_off图标)到AppBar
   - 添加相应的i18n字符串
6. api_test_screen.dart - 修复硬编码英文 'API Test' (改用loc.apiTest)
7. i18n/app_localizations.dart - 添加showHiddenFiles和hideHiddenFiles字符串

+++ lib/ui/viewer/widgets/video_player_widget.dart
@@ -162,9 +162,8 @@ class _VideoPlayerWidgetState extends State<VideoPlayerWidget> {
     final position = controller.value.position;
     final duration = controller.value.duration;
     final maxMs =
-        duration.inMilliseconds.toDouble().clamp(1, double.infinity).toDouble();
-    final valueMs =
-        position.inMilliseconds.toDouble().clamp(0, maxMs).toDouble();
+        duration.inMilliseconds.toDouble().clamp(1.0, double.infinity);
+    final valueMs = position.inMilliseconds.toDouble().clamp(0.0, maxMs);
 
     return GestureDetector(
       onTap: _toggleControls,
